initial refactor to use spine runtimes

+++ core/src/main/dc/targetman/mechanics/VitalLimbsSystem.kt
@@ -1,5 +1,6 @@
 package dc.targetman.mechanics
 
+import dc.targetman.epf.parts.SkeletonPart
 import dc.targetman.epf.parts.VitalLimbsPart
 import dclib.epf.Entity
 import dclib.epf.EntityManager
@@ -11,9 +12,12 @@ class VitalLimbsSystem(entityManager: EntityManager) : EntitySystem(entityManage
 	// TODO: Subscribe to entity removed event instead of polling
 	override fun update(delta: Float, entity: Entity) {
 		val vitalLimbsPart = entity.tryGet(VitalLimbsPart::class.java)
-		val isVitalLimbDead = vitalLimbsPart?.limbs.orEmpty().any { !it.entity.isActive }
-		if (isVitalLimbDead) {
-			entityManager.remove(entity)
+        if (vitalLimbsPart != null) {
+            val skeletonPart = entity[SkeletonPart::class.java]
+            val isVitalLimbDead = vitalLimbsPart.limbNames.any { skeletonPart[it].isActive }
+            if (isVitalLimbDead) {
+                entityManager.remove(entity)
+            }
 		}
 	}
 }
